Añadidas funcionalidades en nuevoPedido

+++ src/main/java/com/cibertec/repository/IProductoRepository.java
@@ -5,6 +5,6 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.cibertec.model.Producto;
 
-public interface IProductoRepository extends JpaRepository<Producto, Long> {
+public interface IProductoRepository extends JpaRepository<Producto, Integer> {
 	List<Producto> findByIdCategoriaProd(int idCategoriaProd);
 }
